- Added a function to serialize objects to JSON - Added a Utilities folder for functions and variables which are going to be used numerous times throughtout the project - Code refactoring
