+++ API/db.json
@@ -1,46 +1,46 @@
 [
-	{
-		"Id": "5a4f00fb-bce7-4901-b026-337606034630",
-		"Type": 1,
-		"Name": "Domas",
-		"Surname": "Nemanius",
-		"Email": "domas@nemanius.com",
-		"PhoneNumber": "+37063666660",
-		"Address": null,
-		"Car": {
-			"FuelConsumption": 5.4,
-			"Make": "Wolkswagen Passat 1986",
-			"RegistrationNumber": "ABC-123"
-		}
-	},
-	{
-		"Id": "5330265b-1668-42c2-91d2-a963da01f144",
-		"Type": 0,
-		"Name": "Adomas",
-		"Surname": "Vensas",
-		"Email": "adomas@vensas.com",
-		"PhoneNumber": "+37063666661",
-		"Address": null,
-		"Car": null
-	},
-	{
-		"Id": "ac8e4bda-9159-4a42-acf0-a77053d8a664",
-		"Type": 0,
-		"Name": "Andrius",
-		"Surname": "Paulauskas",
-		"Email": "andrius@paulauskas.com",
-		"PhoneNumber": "+37063666662",
-		"Address": null,
-		"Car": null
-	},
-	{
-		"Id": "f4228c15-c46f-4f85-b586-f7da30738c27",
-		"Type": 0,
-		"Name": "Kamile",
-		"Surname": "Samusiovaite",
-		"Email": "domas@nemanius.com",
-		"PhoneNumber": "+37063666663",
-		"Address": null,
-		"Car": null
-	}
+  {
+    "Id": "5a4f00fb-bce7-4901-b026-337606034630",
+    "Type": 1,
+    "Name": "Domas",
+    "Surname": "Nemanius",
+    "Email": "domas@nemanius.com",
+    "PhoneNumber": "+37063666660",
+    "Address": null,
+    "Car": {
+      "FuelConsumption": 5.4,
+      "Make": "Wolkswagen Passat 1986",
+      "RegistrationNumber": "ABC-123"
+    }
+  },
+  {
+    "Id": "5330265b-1668-42c2-91d2-a963da01f144",
+    "Type": 0,
+    "Name": "Adomas",
+    "Surname": "Vensas",
+    "Email": "adomas@vensas.com",
+    "PhoneNumber": "+37063666661",
+    "Address": null,
+    "Car": null
+  },
+  {
+    "Id": "ac8e4bda-9159-4a42-acf0-a77053d8a664",
+    "Type": 0,
+    "Name": "Andrius",
+    "Surname": "Paulauskas",
+    "Email": "andrius@paulauskas.com",
+    "PhoneNumber": "+37063666662",
+    "Address": null,
+    "Car": null
+  },
+  {
+    "Id": "f4228c15-c46f-4f85-b586-f7da30738c27",
+    "Type": 0,
+    "Name": "Kamile",
+    "Surname": "Samusiovaite",
+    "Email": "domas@nemanius.com",
+    "PhoneNumber": "+37063666663",
+    "Address": null,
+    "Car": null
+  }
 ]
